# server/swagger.json
{"openapi":"3.0.0","paths":{"/login":{"post":{"operationId":"AppController_login","summary":"Login.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"User logged.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessDto"}}}},"401":{"description":"Wrong credentials."}},"tags":["app"]}},"/me":{"get":{"operationId":"AppController_getMe","summary":"Fetch information about token.","parameters":[],"responses":{"200":{"description":"User info.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Invalid token."}},"tags":["app"],"security":[{"bearer":[]}]}},"/nodes":{"get":{"operationId":"AppController_getNodes","summary":"Fetch saved nodes.","parameters":[],"responses":{"200":{"description":"Get nodes.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodesDto"}}}},"401":{"description":"Invalid token."}},"tags":["app"],"security":[{"bearer":[]}]},"post":{"operationId":"AppController_saveNodes","summary":"Save all nodes.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodesDto"}}}},"responses":{"200":{"description":"Nodes were saved.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NodesDto"}}}},"401":{"description":"Invalid token."}},"tags":["app"],"security":[{"bearer":[]}]}}},"info":{"title":"Notum nodes","description":"🚀🌙","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"username":{"type":"string","example":"notum"},"password":{"type":"string","example":"toMoon"}},"required":["username","password"]},"AccessDto":{"type":"object","properties":{"access_token":{"type":"string"}},"required":["access_token"]},"UserDto":{"type":"object","properties":{"sub":{"type":"number","description":"User id"},"username":{"type":"string","description":"Username"},"iat":{"type":"number","description":"iat"},"exp":{"type":"number","description":"iat"}},"required":["sub","username"]},"NodeDto":{"type":"object","properties":{"id":{"type":"number","description":"Node id"},"description":{"type":"string","description":"Optional node's description"},"parentId":{"type":"number","description":"Parent ID (root node doesn't have parent)"},"color":{"type":"string","description":"Optional branch color"}},"required":["id"]},"NodesDto":{"type":"object","properties":{"nodes":{"type":"array","items":{"$ref":"#/components/schemas/NodeDto"}}},"required":["nodes"]}}}}